+++ src/App.tsx
@@ -1,11 +1,17 @@
-import { createBrowserRouter, matchPath, Outlet, RouterProvider, useLocation } from "react-router-dom";
+import {
+  createBrowserRouter,
+  matchPath,
+  Outlet,
+  RouterProvider,
+  useLocation,
+} from "react-router-dom";
 import styled, { ThemeProvider } from "styled-components";
 import { theme } from "@/styles/Theme";
 import GlobalStyle from "@/styles/GlobalStyles";
 import BottomNavBar from "@/components/BottomNavBar";
 import LoginPage from "@/pages/LoginPage/LoginPage";
 import SignUpPage from "@/pages/LoginPage/SignUpPage";
-import HomePage from "@/pages/HomePage";
+import HomePage from "@/pages/HomePage/HomePage";
 import VoiceRoomListPage from "@/pages/VoiceRoomPage/VoiceRoomListPage";
 import ChatPage from "@/pages/ChatPage/ChatPage";
 import ChattingPage from "@/pages/ChatPage/ChattingPage/ChattingPage";
@@ -37,70 +43,72 @@ import EditVoiceRoomPage from "./pages/VoiceRoomPage/EditVoiceRoomPage";
 // };
 
 const LayoutContainer = styled.div`
-	position: relative;
-	min-width: 360px;
-	max-width: 720px;
-	width: 100%;
-	margin: 0 auto;
+  position: relative;
+  min-width: 360px;
+  max-width: 720px;
+  width: 100%;
+  margin: 0 auto;
 
-	#content {
-		min-height: calc(100vh - 3.75rem);
-	}
+  #content {
+    min-height: calc(100vh - 3.75rem);
+  }
 `;
 
 interface RouteChildren {
-	path: string;
-	element: JSX.Element;
-	hasBottomBar?: boolean;
+  path: string;
+  element: JSX.Element;
+  hasBottomBar?: boolean;
 }
 
 function Layout({ routes_children }: { routes_children: RouteChildren[] }) {
-	const { pathname } = useLocation();
+  const { pathname } = useLocation();
 
-	return (
-		<ThemeProvider theme={theme}>
-			<GlobalStyle />
-			<LayoutContainer>
-				<div id="content">
-					<Outlet />
-				</div>
-				{routes_children.find((child) => matchPath(child.path, pathname))?.hasBottomBar && <BottomNavBar />}
-			</LayoutContainer>
-		</ThemeProvider>
-	);
+  return (
+    <ThemeProvider theme={theme}>
+      <GlobalStyle />
+      <LayoutContainer>
+        <div id="content">
+          <Outlet />
+        </div>
+        {routes_children.find((child) => matchPath(child.path, pathname))?.hasBottomBar && (
+          <BottomNavBar />
+        )}
+      </LayoutContainer>
+    </ThemeProvider>
+  );
 }
 
 function App() {
-	const routes_children = [
-		{ path: "/", element: <HomePage />, hasBottomBar: true },
-		{ path: "/voiceroom", element: <VoiceRoomListPage />, hasBottomBar: true },
-		{ path: "/chat", element: <ChatPage />, hasBottomBar: true },
-		{ path: "/chat/:id", element: <ChattingPage /> },
-		{ path: "/chat/create", element: <ChatCreatePage /> },
-		{ path: "/pay", element: <PayPage />, hasBottomBar: true },
-		{ path: "/board", element: <BoardPage />, hasBottomBar: true },
-		{ path: "/space", element: <SpacePage /> },
-		{ path: "/space/addspace", element: <AddSpacePage /> },
-		{ path: "/space/spaceoption", element: <SpaceOption /> },
-		{ path: "/space/spaceoption/accountmanage", element: <AccountManage /> },
-		{ path: "/space/spaceoption/profilemanage", element: <ProfileManage /> },
-		{ path: "/space/spaceoption/alarmmanage", element: <AlarmManage /> },
-		{ path: "/createvoiceroom", element: <CreateVoiceRoomPage />, hasBottomBar: false },
-		{ path: "/joinvoiceroom", element: <JoinVoiceRoomPage />, hasBottombar: false },
-		{ path: "/editvoiceroom", element: <EditVoiceRoomPage />, hasBottombar: false },
-		{ path: "/login", element: <LoginPage />, hasBottombar: false },
-		{ path: "/signup", element: <SignUpPage />, hasBottombar: false },
-	];
+  const routes_children = [
+    { path: "/", element: <HomePage />, hasBottomBar: true },
+    { path: "/voiceroom", element: <VoiceRoomListPage />, hasBottomBar: true },
+    { path: "/chat", element: <ChatPage />, hasBottomBar: true },
+    { path: "/chat/:id", element: <ChattingPage /> },
+    { path: "/chat/create", element: <ChatCreatePage /> },
+    { path: "/pay", element: <PayPage />, hasBottomBar: true },
+    { path: "/board", element: <BoardPage />, hasBottomBar: true },
+    { path: "/space", element: <SpacePage /> },
+    { path: "/space/addspace", element: <AddSpacePage /> },
+    { path: "/space/spaceoption", element: <SpaceOption /> },
+    { path: "/space/spaceoption/accountmanage", element: <AccountManage /> },
+    { path: "/space/spaceoption/profilemanage", element: <ProfileManage /> },
+    { path: "/space/spaceoption/alarmmanage", element: <AlarmManage /> },
+    { path: "/createvoiceroom", element: <CreateVoiceRoomPage />, hasBottomBar: false },
+    { path: "/joinvoiceroom", element: <JoinVoiceRoomPage />, hasBottombar: false },
+    { path: "/editvoiceroom", element: <EditVoiceRoomPage />, hasBottombar: false },
+    { path: "/login", element: <LoginPage />, hasBottombar: false },
+    { path: "/signup", element: <SignUpPage />, hasBottombar: false },
+  ];
 
-	const routes = [
-		{
-			element: <Layout routes_children={routes_children} />,
-			children: routes_children,
-		},
-	];
-	const router = createBrowserRouter(routes);
+  const routes = [
+    {
+      element: <Layout routes_children={routes_children} />,
+      children: routes_children,
+    },
+  ];
+  const router = createBrowserRouter(routes);
 
-	return <RouterProvider router={router} />;
+  return <RouterProvider router={router} />;
 }
 
 export default App;
